refactor: extract async test utilities to shared test_helpers module
Create a shared test_helpers module to eliminate code duplication of
async test functions across different test modules. This enables
better code reuse and consistency in async device testing.

Changes:
- Add src/test_helpers.rs with shared async test utilities:
  * io_async_fn() - async I/O function for null device testing
  * q_async_fn() - queue async function for spawning I/O tasks
  * device_handler_async() - complete async device handler
  * ublk_join_tasks() - utility for managing task execution
- Update lib.rs to include test_helpers module with cfg(test)
- Refactor ctrl_async.rs to use shared functions instead of local copies
- Add example usage in io.rs demonstrating cross-module sharing
- All existing tests continue to work unchanged

Benefits:
- Eliminates duplication of complex async test logic
- Provides consistent test patterns across modules
- Enables easy addition of new shared test utilities
- Improves maintainability with single source of truth
